:construction: line chart rebuild
